% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrowthExperiment.R
\name{GrowthExperiment-coerce}
\alias{GrowthExperiment-coerce}
\title{Coerce to a \code{GrowthExperiment}}
\arguments{
\item{from}{A \code{data.frame}, \code{matrix}, or \code{SummarizedExperiment} to coerce.}
}
\value{
A \link[=GrowthExperiment-class]{GrowthExperiment} object.
}
\description{
Standard S4 coercion methods for \link[=GrowthExperiment-class]{GrowthExperiment}. Coercion in the
opposite direction, \code{as(x, "SummarizedExperiment")}, is inherited from the
class hierarchy and always available.
}
\examples{
data(yeast_growth_data)

ge <- as(yeast_growth_data, "GrowthExperiment")
ge

se <- as(ge, "SummarizedExperiment")
as(se, "GrowthExperiment")
}
