% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/GrowthFit-methods.R
\name{GrowthFit-accessors}
\alias{GrowthFit-accessors}
\alias{fit_sample}
\alias{fit_model}
\alias{fit_status}
\alias{fit_converged}
\alias{fit_data}
\alias{fit_sample,GrowthFit-method}
\alias{fit_model,GrowthFit-method}
\alias{fit_status,GrowthFit-method}
\alias{fit_converged,GrowthFit-method}
\alias{fit_data,GrowthFit-method}
\title{Accessors for \code{GrowthFit} Objects}
\usage{
fit_sample(x, ...)

fit_model(x, ...)

fit_status(x, ...)

fit_converged(x, ...)

fit_data(x, ...)

\S4method{fit_sample}{GrowthFit}(x, ...)

\S4method{fit_model}{GrowthFit}(x, ...)

\S4method{fit_status}{GrowthFit}(x, ...)

\S4method{fit_converged}{GrowthFit}(x, ...)

\S4method{fit_data}{GrowthFit}(x, ...)
}
\arguments{
\item{x}{A \link[=GrowthFit-class]{GrowthFit} object.}

\item{...}{Arguments passed to methods.}
}
\value{
\itemize{
\item \code{fit_sample()}: a character scalar naming the fitted sample.
\item \code{fit_model()}: a character scalar naming the fitted model.
\item \code{fit_status()}: a character scalar describing the fit outcome.
\item \code{fit_converged()}: a logical scalar.
\item \code{fit_data()}: a tibble of the observed data used for fitting.
}
}
\description{
Extract individual components of a \link[=GrowthFit-class]{GrowthFit} object. These
accessors are the supported interface; slots should not be accessed
directly with \code{@}.
}
\examples{
data(yeast_growth_data)
tidy_growth <- as_tidy_growth_data(yeast_growth_data)
sample_id <- unique(tidy_growth$sample)[1]
fit <- fit_growth_curve(dplyr::filter(tidy_growth, sample == sample_id))

fit_sample(fit)
fit_model(fit)
fit_status(fit)
fit_converged(fit)
head(fit_data(fit))
}
