% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_experiment.R
\name{sample_data}
\alias{sample_data}
\title{Sample Metadata}
\usage{
sample_data(x)
}
\arguments{
\item{x}{A \code{SummarizedExperiment} compatible with \code{growkar}.}
}
\value{
A tibble containing \code{colData(x)}.
}
\description{
Access sample-level metadata from a \code{SummarizedExperiment}.
}
\examples{
data(yeast_growth_data)
se <- GrowthExperiment(yeast_growth_data)
sample_data(se)
}
