% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrowthFit-methods.R
\name{show,GrowthFit-method}
\alias{show,GrowthFit-method}
\title{Display a \code{GrowthFit} Object}
\usage{
\S4method{show}{GrowthFit}(object)
}
\arguments{
\item{object}{A \link[=GrowthFit-class]{GrowthFit} object.}
}
\value{
\code{object}, invisibly. Called for its side effect of printing.
}
\description{
Compact one-line summary of a fitted growth model.
}
\examples{
data(yeast_growth_data)
tidy_growth <- as_tidy_growth_data(yeast_growth_data)
sample_id <- unique(tidy_growth$sample)[1]
fit <- fit_growth_curve(dplyr::filter(tidy_growth, sample == sample_id))
fit
}
