% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_growth_fit.R
\name{augment_growth_fit}
\alias{augment_growth_fit}
\title{Augment Observed Data with Fitted Values}
\usage{
augment_growth_fit(fit)
}
\arguments{
\item{fit}{A \code{growkar_fit} object.}
}
\value{
A tidy tibble containing observed and fitted values.
}
\description{
Return observed data together with fitted values from a \code{growkar_fit} object.
}
