% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_growth_rate.R
\name{calculate_growth_rate}
\alias{calculate_growth_rate}
\title{calculate_growth_rate}
\usage{
calculate_growth_rate(
  dat_growth_curve,
  average_replicates = FALSE,
  first_timepoint = 0,
  select_replicate = NULL,
  cells_at_OD_1 = 3e+07
)
}
\arguments{
\item{dat_growth_curve}{a tibble, with first column as `Time` and after that
columns as samples. If the samples have replicates mention them as
`sampleName_replicate`, for example `ScPH4_R1` , `ScPH4_R2`, `ScPH4_R3`}

\item{average_replicates}{logical, to plot average and standard deviation of
replicates, Default: FALSE}

\item{first_timepoint}{numeric, first time-point from where to calculate the
lag-phase;, Default: 0}

\item{select_replicate}{a vector, define replicates to be selected to plot,
it should be anything after `_`. for eg. if only replicate R1 and R2 need
to be plotted from samples ScPH4_R1, ScPH4_R2; the vector should be
`c("R1", "R2")`, Default: NULL}

\item{cells_at_OD_1}{numeric, a mutplication factor to use absolute cell
number at OD=1, for instance OD600=1 represents 3e+07 yeast cells. Default:
3e+07}
}
\value{
a table with growth rate (generations/hour) and generation time
  (minutes/generation)
}
\description{
calculates growth rate and generation time by computing the
  log-phase start and end time.
}
\details{
Start and End-point are computed as; Start time-point: is time at which OD is appx
  double of the first-time point and end time-point: is time from which stationary
  phase is beginning.
}
\examples{
\dontrun{
if(interactive()){
 calculate_growth_rate(dat_growth_curve, average_replicates = TRUE, select_replicate = c("R1", "R2", "R4"))
 }
}
}
\seealso{
\code{\link[tidyr]{gather}},\code{\link[tidyr]{separate}}
  \code{\link[dplyr]{mutate}},\code{\link[dplyr]{filter}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{select}},\code{\link[dplyr]{join}},\code{\link[dplyr]{slice}},\code{\link[dplyr]{summarise}},\code{\link[dplyr]{bind}}
   \code{\link[forcats]{as_factor}}
}
