% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_params.R
\name{extract_params}
\alias{extract_params}
\title{Extract Fitted Parameters}
\usage{
extract_params(fit)
}
\arguments{
\item{fit}{A \code{growkar_fit} object returned by \code{fit_growth_curve()}.}
}
\value{
A tibble of fitted parameters.
}
\description{
Extract coefficients and derived doubling time from a \code{growkar_fit} object.
}
