% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growkar-package.R
\docType{package}
\name{growkar-package}
\alias{growkar}
\alias{growkar-package}
\title{growkar: Tidy Analysis of Microbial Growth Curves}
\description{
Internal package-level declarations for shared imports used throughout the
tidy-v2 implementation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sethiyap/growkar}
  \item Report bugs at \url{https://github.com/sethiyap/growkar/issues}
}

}
\author{
\strong{Maintainer}: Pooja Sethiya \email{pooja.sethiya@wimr.org.au}

}
\keyword{internal}
