% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growkar-package.R
\docType{package}
\name{growkar-package}
\alias{growkar}
\alias{growkar-package}
\title{growkar: Tidy Analysis of Microbial Growth Curves}
\description{
Internal package-level declarations for shared imports used throughout the
current tidy and Bioconductor-interoperable implementation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sethiyap/growkar}
  \item Report bugs at \url{https://github.com/sethiyap/growkar/issues}
}

}
\author{
\strong{Maintainer}: Pooja Sethiya \email{poojasethiya24@gmail.com} (\href{https://orcid.org/0000-0001-6584-3912}{ORCID})

}
\keyword{internal}
