% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_experiment.R
\name{growth_assay}
\alias{growth_assay}
\title{Growth Assay Matrix}
\usage{
growth_assay(x)
}
\arguments{
\item{x}{A \code{SummarizedExperiment} compatible with \code{growkar}.}
}
\value{
A numeric matrix with time points in rows and samples in columns.
}
\description{
Access the canonical OD assay from a \code{SummarizedExperiment}.
}
\examples{
data(yeast_growth_data)
se <- as_summarized_experiment(yeast_growth_data)
growth_assay(se)
}
