% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fitted_curve.R
\name{plot_fitted_curve}
\alias{plot_fitted_curve}
\title{Plot a Fitted Growth Curve}
\usage{
plot_fitted_curve(fit)
}
\arguments{
\item{fit}{A \code{growkar_fit} object.}
}
\value{
A \code{ggplot2} object.
}
\description{
Overlay observed points and fitted values from a parametric growth model.
}
