% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_growth_curve.R
\name{plot_growth_curve}
\alias{plot_growth_curve}
\title{plot_growth_curve}
\usage{
plot_growth_curve(
  dat_growth_curve,
  average_replicates = FALSE,
  select_replicate = NULL
)
}
\arguments{
\item{dat_growth_curve}{a tibble, with first column as `Time` and after that
columns as samples. If the samples have replicates mention them as
`sampleName_replicate`, for example `ScPH4_R1` , `ScPH4_R2`, `ScPH4_R3`}

\item{average_replicates}{logical, to plot average and standard deviation of
replicates}

\item{select_replicate}{a vector, define replicates to be selected to plot,
it should be anything after `_`. for eg. if only replicate R1 and R2 need
to be plotted from samples ScPH4_R1, ScPH4_R2; the vector should be
`c("R1", "R2")`; default: NULL}
}
\value{
a plot of growth curve
}
\description{
plot_growth_curve
}
\examples{

\dontrun{

dat_growth_curve <- readr::read_delim(pipe("pbpaste"), delim="\t",col_names = TRUE)

plot_growth_curve(dat_growth_curve = "data/growth_curve_data.rda")

}

}
