% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_growth_curve.R
\name{plot_growth_curve}
\alias{plot_growth_curve}
\title{plot_growth_curve}
\usage{
plot_growth_curve(
  dat_growth_curve,
  average_replicates = FALSE,
  select_replicates = NULL
)
}
\arguments{
\item{dat_growth_curve}{a tibble, with first column as `Time` and after that
columns as samples. If the samples have replicates mention them as
`sampleName_replicate`, for example `CgFlu_R1` , `CgFlu_R2`, `CgFlu_R3`}

\item{average_replicates}{logical, to plot average and standard deviation of
replicates, Default: FALSE}

\item{select_replicate}{a vector, define replicates to be selected to plot,
it should be anything after `_`. for eg. if only replicate R1 and R2 need
to be plotted from samples CgFlu_R1, CgFlu_R2; the vector should be
`c("R1", "R2")`, Default: NULL}
}
\value{
a plot of growth curve
}
\description{
Visualization of growth curve from Optical Density values generated by spectrophotometer like Cytation3
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){

  plot_growth_curve(dat_growth_curve = yeast_growth_data,average_replicates=TRUE, select_replicates=c("R1", "R2"))

  }
}
}
\seealso{
\code{\link[tidyr]{gather}},\code{\link[tidyr]{separate}}
 \code{\link[dplyr]{filter}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{select}}
 \code{\link[forcats]{as_factor}}
 \code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{geom_point}},\code{\link[ggplot2]{geom_path}},\code{\link[ggplot2]{geom_crossbar}},\code{\link[ggplot2]{position_dodge}},\code{\link[ggplot2]{labs}},\code{\link[ggplot2]{facet_grid}},\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{margin}},\code{\link[ggplot2]{ggtheme}},\code{\link[ggplot2]{scale_manual}}
 \code{\link[RColorBrewer]{RColorBrewer}}
}
