% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_palette.R
\name{select_palette}
\alias{select_palette}
\title{select_palette}
\usage{
select_palette(num_of_colors, palette_name = "all_colors")
}
\arguments{
\item{num_of_colors}{numeric, number of colors to be obtained from palette}

\item{palette_name}{character,name palettes Default: 'all_colors'
For individual palette;
 \itemize{
 \item {Accent: }{8}
 \item {Dark2: }{8}
 \item {Paired: }{12}
 \item {Pastel1: }{9}
 \item {Pastel2: }{8}
 \item {Set1: }{9}
 \item {Set2: }{8}
 \item {Set3: }{12} }}
}
\value{
a vector of colors from selected palette
}
\description{
allows to select contrasting colors from 8 different palettes
  contributing a total of 75 colors
}
\details{
select_palette is particularly useful as individual palette from
  many color brewer packages contain less than 10 colors
}
\examples{
\dontrun{
if(interactive()){
 # colors combined from all palettes
 select_palette(num_of_colors = 10)

 # colors from Accent palette
 select_palette(num_of_colors = 4, palette_name = "Accent")
 }
}
}
\keyword{internal}
