% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_experiment.R
\name{timepoints}
\alias{timepoints}
\title{Timepoint Metadata}
\usage{
timepoints(x)
}
\arguments{
\item{x}{A \code{SummarizedExperiment} compatible with \code{growkar}.}
}
\value{
A tibble containing \code{rowData(x)}.
}
\description{
Access timepoint metadata from a \code{SummarizedExperiment}.
}
\examples{
data(yeast_growth_data)
se <- as_summarized_experiment(yeast_growth_data)
timepoints(se)
}
